requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
ratelimit>=2.2.1
python-dateutil>=2.8.2
loguru>=0.7.0
